+++ src/components/NavLinks.svelte
@@ -4,6 +4,7 @@
 </script>
 
 <Link to="/about" class="link">About</Link>
-<Link to="/contact" class="link">Contact</Link>
+<!-- <Link to="/contact" class="link">Contact</Link> -->
 <Link to="/blog" class="link">Blog</Link>
-<Link to="/fullpage" class="link">Full Page</Link>
+<Link to="/changelog" class="link">Changelog</Link>
+<Link to="/fullpage" class="link">FP Demo</Link>